+++ BallBalancing/Environment.py
@@ -1,6 +1,23 @@
 import numpy as np
 
 
+def generateQTable():
+    """
+    Create qtable for the agent
+    :return: QTable
+    """
+    dict = {}
+    positions = np.round(list(np.linspace(-1, 1, 100)), decimals=1)
+    velocities = np.round(list(np.linspace(-3, 3, 50)), decimals=1)
+    actions = np.round(list(np.linspace(-1, 1, 15)), decimals=1)
+    for p in positions:
+        for v in velocities:
+            dict[(p, v)] = {}
+            for a in actions:
+                dict[(p, v)][a] = 0.0
+    return dict
+
+
 def reward(x, v):
     """
     Reward function
@@ -84,5 +101,3 @@ def choose_action(states, actions, qTables, epsilon=0.2):
             action = max(qTables[q][states].values())
         new_actions[q] = action
     return new_actions
-
-
