+++ BallBalancing/Training.py
@@ -1,6 +1,8 @@
 from QLearningAlgorithms import *
 from Utility import *
 import pandas as pd
+import matplotlib.pyplot as plt
+from BallBalancing.Environment import *
 
 
 def trainDistributed():
@@ -11,8 +13,10 @@ def trainDistributed():
     qTable1 = generateQTable()
     qTable2 = generateQTable()
     qTables = [qTable1, qTable2]
+
     actions = np.round(np.linspace(-1, 1, 15), decimals=1)
     output = []
+
     for trial in range(5000):
         states = (0.5, 0.1)  # initial states
         rewardSum = 0
@@ -25,6 +29,7 @@ def trainDistributed():
 
             r = reward(x, v)
             rewardSum = rewardSum + r
+
             states = (np.round(x, decimals=1), np.round(v, decimals=1))
             states = check_states(states, qTables)  # check if the states have a match in the discrete grid
 
@@ -32,6 +37,7 @@ def trainDistributed():
             new_actions = check_actions(new_actions, actions)
 
             qTables = distributed(qTables, r, states, new_actions, alpha)
+
         output.append(rewardSum)
         # if trial == 0:
         #     output.append(rewardSum)
@@ -55,8 +61,10 @@ def trainDecentralized():
     qTable1 = generateQTable()
     qTable2 = generateQTable()
     qTables = [qTable1, qTable2]
+
     actions = np.round(np.linspace(-1, 1, 15), decimals=1)
     output = []
+
     for trial in range(5000):
         states = (0.5, 0.1)  # initial states
         rewardSum = 0
@@ -69,6 +77,7 @@ def trainDecentralized():
 
             r = reward(x, v)
             rewardSum = rewardSum + r
+
             new_states = (np.round(x, decimals=1), np.round(v, decimals=1))
             new_states = check_states(new_states, qTables)  # check if the states have a match in the discrete grid
 
@@ -77,14 +86,16 @@ def trainDecentralized():
 
             qTables = decentralized(qTables, states, new_actions, alpha, r, gamma, new_states)
             states = new_states
+
         output.append(rewardSum)
         # if trial == 0:
         #     output.append(rewardSum)
         # else:
         #     output.append(output[trial-1]+rewardSum)
+
     plt.scatter(list(range(5000)), output)
     plt.show()
-    print(qTables[0])
+
     countNot0(qTables)
 
     pd.DataFrame.from_dict(qTables[0], orient='index').to_csv('./QTables/qT1_Decentralized.csv')
@@ -101,8 +112,10 @@ def trainHysteretic():
     qTable1 = generateQTable()
     qTable2 = generateQTable()
     qTables = [qTable1, qTable2]
+
     actions = np.round(np.linspace(-1, 1, 15), decimals=1)
     output = []
+
     for trial in range(5000):
         states = (0.5, 0.1)  # initial states
         rewardSum = 0
@@ -115,6 +128,7 @@ def trainHysteretic():
 
             r = reward(states[0], states[1])
             rewardSum = rewardSum + r
+
             new_states = (np.round(x, decimals=1), np.round(v, decimals=1))
             new_states = check_states(new_states, qTables)  # check if the states have a match in the discrete grid
 
@@ -123,6 +137,7 @@ def trainHysteretic():
 
             qTables = hysteretic(qTables, states, new_actions, alpha, beta, r, gamma, new_states)
             states = new_states
+
         output.append(rewardSum)
         # if trial == 0:
         #     output.append(rewardSum)
@@ -130,12 +145,12 @@ def trainHysteretic():
         #     output.append(output[trial-1]+rewardSum)
     plt.scatter(list(range(5000)), output)
     plt.show()
-    print(len(output))
-    print(qTables[0])
+
     countNot0(qTables)
 
     pd.DataFrame.from_dict(qTables[0], orient='index').to_csv('./QTables/qT1_Hysteretic.csv')
     pd.DataFrame.from_dict(qTables[1], orient='index').to_csv('./QTables/qT2_Hysteretic.csv')
+
 
 print('Distributed case')
 trainDistributed()
