+++ TwoLinkRigidManipulator/EnvironmentRigidManipulator.py
@@ -1,0 +1,41 @@
+import numpy as np
+
+g = 9.81
+l1 = l2 = 0.1
+m1 = 1.25
+m2 = 1
+i1 = 0.004
+i2 = 0.003
+c1 = c2 = 0.05
+b1 = 0.1
+b2 = 0.02
+p1 = m1 * c1 * c1 + m2 * l1 * l1 + i1
+p2 = m2 * c2 * c2 + i2
+p3 = m2 * l1 * c2
+g1 = (m1 * c1 + m2 * c1) * g
+g2 = m2 * c2 * g
+
+
+def massMatrix(theta):
+    m = np.zeros(shape=(2, 2))
+    m[0][0] = p1 + p2 + p3 * (np.cos(theta[1]))
+    m[0][1] = p2 + p3 * (np.cos(theta[1]))
+    m[1][0] = p2 + p3 * (np.cos(theta[1]))
+    m[1][1] = p2
+    return m
+
+
+def coriolisMatrix(theta, theta_point):
+    m = np.zeros(shape=(2, 2))
+    m[0][0] = b1 - p3 * theta_point[1] * (np.sin(theta[1]))
+    m[0][1] = - p3 * (theta_point[0] + theta_point[1]) * (np.sin(theta[1]))
+    m[1][0] = - p3 * (theta_point[0] + theta_point[1]) * (np.sin(theta[1]))
+    m[1][1] = b2
+    return m
+
+
+def centrifugalMatrix(theta):
+    m = np.zeros(shape=(2, 1))
+    m[0][0] = -g1 * np.sin(theta[0]) - g2 * np.sin((theta[0] + theta[1]))
+    m[1][0] = -g2 * np.sin((theta[0] + theta[1]))
+    return m
